Resolve some issues with lazy loading
Altough on_file_load approach works on the home laptop, it does not work
on the office one. The problem is that the packer_compiled.lua file is
created differently on the two machines.

The main culprit seems to be on_file_load. I have removed it and made
most of the commands lazy loaded, as much as possible.