Update NVIM config
Main desire was update my nvim config to use Lazy package manager
instead of Packer (since latter has been unmaintained for a while).

This change brought with it a wave of changes in the config directory
file structure:
* Plugins are now contained in the plugins folder. The contents of each
  file follows the spec defined by the Lazy package manager, which can
  automatically load all plugins, just by pointing the setup function to
  the plugins directory.
* Some of the plugins were removed, some where added
* I dropped null-ls, since it is unmaintained, I am using conform for
  formatting and nvim-lint for linting.
* I finally added which-key, I wanted to do this for a while.
* I generally cleaned up configuration each plugin. LSP folder is now a
  bit nicer.

My main source of the inspiration for the bootstrapping and the
configuration structure was NVChad and LazyVim.

Unfinished work:
* Gitlint requires descriptions for the mappings.
* colorscheme.lua only supports tokyonight for now.