Transition from i3 to Waybar
Entire setup was overhauled for Waybar. Although the script for i3 still
exists, it didn't work last time.

Most of the i3 specific configs weren't deleted, they just aren't linked
to the home directory anymore.

The system is almost stable.

It is mostly missing the theme support for Wallust, gtk themes and support
for Manjaro installation.

Only Ubuntu Server is supported for now, but not yet fully tested.

Wallpapers were moved to a new repository.